requirements.txt
streamlit==1.29.0
pandas==2.1.3
Pillow==10.1.0
selenium==4.15.2
beautifulsoup4==4.12.2
